requirements.txt
asgiref==3.6.0
certifi==2022.12.7

celery==5.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.0


Django==4.2
django-ckeditor==6.5.1
django-filter==23.1
django-restframework==0.0.1
djangorestframework==3.14.0

environ==1.0

Pillow==9.5.0
psycopg==3.1.8
python-dateutil==2.8.2
python-decouple==3.8

postgis==1.0.4

redis==4.5.4
async-timeout==4.0.2

Token==1.8.6

xlwt==1.3.0